Add files via upload
These are the redesigned webpages. All of the buttons and links work except for the sign-up/in, which I will handle soon. The Account Sign Up page is set up for now and will be finished once connected to backend. To summarize, we should use this as a base because everything currently works.